Created new functions and changed code layout
The new functions were created to clear the grid content by removing
the row div and setting the grid event listeners for each tile. The
clear function will be activated every time the button to set the grid
is clicked. Since the clear function removes the row divs, the event
listeners must be added to the new row divs created after the user input
is taken in. This is why the setTileListener function was made, which
takes in the variable that stores a node of query selector for the tile
classes.