Fixed issue cased by misnamed function
An issue occurred because the setTileListener function was renamed to
setTileEvent. This name changed was not implemented across the pieces
of code that uses this function. This resulted in the event listener
failing to activate whenever the mouse moved across the grid. The issue
has been resolved now that the any reference to to old function name has
been  updated.