Created new function to assign a random tile color
The function, called getRandomRgbColor, returns a string that will be
used with the event listener inside the setTileEvent function to give
that tile a random background color. An if statement inside the
setTileEvent functions prevents the tile from being assigned a new color
if it has already been given one. This is achieved by adding a class
named "changed" after a random color is assigned. if this class name
doesn't exist, then the statement  gives the tile div the class name and
changes the background color.