Completed darkenRgbColor function
This function takes in a tile div as a parameter. It gets the background
color, creates a substring that is then split. The values are then
assigned to three different variables, which are used to reduce each RGB
value by 10%. This decrease doesn't occur if the result would reduce the
value below 0. The result is then used to return a new RGB string.