Fixed setUserDimension function and event listener
The setUserDimension function caused an issue where clicking the cancel
button on the prompt wouldn't close the prompt. An if statement was
added to check if the result is null, which is the value returned when
the cancel button is clicked. This fix caused an issue with the event
listener for the button that sets a new grid based on the previously
mentioned function. Clicking cancel would delete the entire grid. This
issue was solved by moving the gridDimension to the top of the listener
and adding a if statement that prevents the other functions from running
if the value is null.