Made some alterations to margins and colors
Made some adjustments to improve the presentation of the page. The
container div was given a grey border. The background color of the tile
was removed, as it was given the color red for testing purposes. The
tile div was given a light grey border. The margins for the h1 and
button tags were also adjusted slightly.