Fixed issue caused by setUserDimension function
The nature in which the gridCount variable was assigned, along with
the structure of the while loop, cause an issue in which the cancel
button from the prompt would not work after the ok button was clicked
with no value entered. The cancel button would not close the prompt.
Furthermore, if a number was entered after the failed cancel attempt,
the grid tiles would be removed completely. To fix this, the gridCount
variable is initialized with an empty string, the result of the prompt
is not immediately cast into a number, the loop contains a break if the
current value is null, meaning that the cancel button was clicked, and
an if condition is used to activate the alert in the function.